# Upgrade to go 1.22.1 (with fix for dockerfiles)