--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -12,7 +12,7 @@ jobs:
   linux-tests:
     strategy:
       matrix:
-        go: [ 1.21.x ]
+        go: [ 1.22.x ]
     runs-on: ubuntu-20.04
     timeout-minutes: 15
 
@@ -40,7 +40,7 @@ jobs:
     - name: Setup Go
       uses: actions/setup-go@v3
       with:
-        go-version: "1.21"
+        go-version: "1.22"
     - name: checkout code
       uses: actions/checkout@v3
     - name: golangci-lint