--- a/tools/package_gcsfuse_docker/Dockerfile
+++ b/tools/package_gcsfuse_docker/Dockerfile
@@ -30,7 +30,7 @@ ARG GCSFUSE_REPO="https://github.com/googlecloudplatform/gcsfuse/"
 ARG BRANCH_NAME="v${GCSFUSE_VERSION}"
 RUN git clone --branch ${BRANCH_NAME} ${GCSFUSE_REPO}
 
-ARG GCSFUSE_PATH=/go/gcsfuse
+ARG GCSFUSE_PATH=${GOPATH}/gcsfuse
 WORKDIR ${GCSFUSE_PATH}
 
 ARG DEBEMAIL="gcs-fuse-maintainers@google.com"