### Description
* Reapply "Upgrade to golang 1.22 (#1753)" (#1766)
    This reverts commit a63a46fdbecbcfa58591e76a9bf49af6239ae7f1.
* Fixes dockerfiles which failed with go 1.22, because of `go get` error, by switching to `git clone` (in the docker container).
* Upgrade to 1.22.1

### Link to the issue in case of a bug fix.
NA

### Testing details
1. Manual - Sanity testing
2. Unit tests - NA
3. Integration tests - [Passed](http://fusion2/ci/kokoro/prod%3Agcsfuse%2Fgcp_ubuntu%2Fpresubmits%2Fperf_tests%2Fpresubmit/activity/a557edef-b441-4b82-a37a-06757750634a).
4. Perf tests
```
+--------+------------+--------------+------------+--------------+--------------+
| Branch | File Size  |   Read BW    |  Write BW  | RandRead BW  | RandWrite BW |
+--------+------------+--------------+------------+--------------+--------------+
| Master |  0.25MiB   | 536.37MiB/s  | 1.26MiB/s  |  77.95MiB/s  |  1.08MiB/s   |
|   PR   |  0.25MiB   | 518.86MiB/s  |  1.3MiB/s  |  70.12MiB/s  |  1.31MiB/s   |
|        |            |              |            |              |              |
| Master | 48.828MiB  | 4645.71MiB/s | 80.55MiB/s | 1238.05MiB/s |  79.37MiB/s  |
|   PR   | 48.828MiB  | 4477.76MiB/s | 79.18MiB/s | 1277.78MiB/s |  80.71MiB/s  |
|        |            |              |            |              |              |
| Master | 976.562MiB | 4406.04MiB/s | 33.37MiB/s | 477.44MiB/s  |  32.95MiB/s  |
|   PR   | 976.562MiB | 4423.47MiB/s | 33.91MiB/s | 608.93MiB/s  |  31.9MiB/s   |
|        |            |              |            |              |              |
+--------+------------+--------------+------------+--------------+--------------+
```