@gargnitingoogle 2024-03-11
> * We should sanity it by running it with louhi test-flow.

Ran it once and build went through. I have given it another run after adding the `--depth 1` flag, to reduce cloning time.

> * Also, should check the impact of these changes with GKE team.

Will check with @songjiaxun .