@gargnitingoogle 2024-03-12
@raj-prince 

> > * We should sanity it by running it with louhi test-flow.
> 
> Ran it once and build went through. I have given it another run after adding the `--depth 1` flag, to reduce cloning time.

I checked Louhi test-flow with `--depth 1` and that worked too.

> 
> > * Also, should check the impact of these changes with GKE team.
> 
> Will check with @songjiaxun .

@songjiaxun shared over chat that "GKE Louhi pipeline and the testgrid uses tools/package_gcsfuse_docker/Dockerfile to build gcsfuse binary, so as long as the Dockerfile works, out pipeline should be fine.". So I think we're good there.